renamed all of the flow duration components to be correct
