+++ client/src/components/streamflow/MonthlyFlowStatistics.vue
@@ -1,39 +1,45 @@
-
 <template>
-    <h3>Total Runoff</h3>
-    <div class="date-selectors">
-        <q-select 
-            class="selector"
-            label="Year From"
-            dense
-        />
-        <div class="q-mx-sm">
-            -
-        </div>
-        <q-select 
-            class="selector q-mx-sm"
-            label="Year to"
-            dense
-        />
-        <q-select 
-            class="selector q-mx-sm"
-            label="Month"
-            dense
-        />
-        <q-btn 
-            class="text-bold q-mx-sm"
-            label="reset dates"
-            flat
-            color="primary"
-        />
-    </div>
-    <div class="monthly-flow-container">
-        <div id="monthly-flow-chart-container">
-            <div class="svg-wrap-mf">
-                <svg class="d3-chart-mf">
+    <h3>Monthly Flow Statistics</h3>
+    <div class="flow-duration-container">
+        <div id="flow-duration-chart-container">
+            <div class="svg-wrap-fd">
+                <svg class="d3-chart-fd">
                     <!-- d3 chart content renders here -->
                 </svg>
             </div>
+        </div>
+
+            
+        <div 
+            v-if="showTooltip"
+            class="flow-duration-tooltip"
+            :style="`left: ${tooltipPosition[0]}px; top: ${tooltipPosition[1]}px`"
+        >
+            <q-card>
+                <div 
+                    v-for="(key, idx) in Object.keys(tooltipData)"
+                >
+                    <div
+                        v-if="idx === 0"
+                    >
+                        <div 
+                            class="tooltip-header"
+                        >
+                            <span class="text-h6">{{ tooltipData[key] }}</span>
+                            <div>
+                                Discharge (m³/s)
+                            </div>
+                        </div>
+                    </div>
+                    <div
+                        v-else
+                        class="tooltip-row" 
+                        :class="['Max', 'Median', 'Min'].includes(key) ? 'val' : 'box-val'"
+                    >
+                        {{ key }}: {{ tooltipData[key].toFixed(2) }}
+                    </div>
+                </div>
+            </q-card>
         </div>
     </div>
 </template>
@@ -41,25 +47,22 @@
 <script setup>
 import * as d3 from "d3";
 import { monthAbbrList } from '@/constants/dateHelpers.js';
-import { onMounted, ref, watch } from 'vue';
-
-const emit = defineEmits(['year-range-selected']);
+import { onMounted, ref } from 'vue';
 
 const props = defineProps({
-    data: {
+    startEndYears: {
         type: Array,
         default: () => [],
     },
-    startEndMonths: {
+    data: {
         type: Array,
-        default: () => ['Jan', 'Dec'],
-    },
-});
+        default: () => [],
+    }
+})
 
 const monthDataArr = ref([]);
 const loading = ref(false);
 const monthPercentiles = ref([]);
-const formattedChartData = ref([]);
 
 // chart variables
 const svgWrap = ref();
@@ -72,130 +75,168 @@ const xGrid = ref();
 const yGrid = ref();
 const xAxis = ref();
 const yAxis = ref();
-const xMax = ref();
-const yearRangeArr = ref([1914, 2022]);
-const dataYears = ref([1914]);
+const yMax = ref();
+const yMin = ref();
 
 // brush functionality
 const brushVar = ref();
 const brushEl = ref();
-const brushedYearStart = ref();
-const brushedYearEnd = ref();
+const brushedStart = ref();
+const brushedEnd = ref();
 
 // chart constants
-const width = 400;
-const height = 1000;
+const width = 600;
+const height = 300;
 const margin = {
-    left: 60,
+    left: 50,
     right: 50,
-    top: 30,
+    top: 10,
     bottom: 50
 };
 
-watch(() => props.startEndMonths, () => {
-    processData(props.data);
-    initializeMonthlyFlowChart();
-})
+// tooltip
+const showTooltip = ref(false);
+const tooltipData = ref();
+const tooltipPosition = ref();
+
+const emit = defineEmits(['range-selected']);
 
 onMounted(() => {
     loading.value = true;
-    initializeMonthlyFlowChart();
+    initializeChart();
     loading.value = false;
 });
 
-const initializeMonthlyFlowChart = () => {
+const initializeChart  = () => {
     if (svg.value) {
-        d3.selectAll('.g-els.mf').remove();
+        d3.selectAll('.g-els.fd').remove();
     }
     processData(props.data);
-    svgWrap.value = document.querySelector('.svg-wrap-mf');
+    svgWrap.value = document.querySelector('.svg-wrap-fd');
     svgEl.value = svgWrap.value.querySelector('svg');
     svg.value = d3.select(svgEl.value)
         .attr("width", width + margin.left + margin.right)
+        .attr("height", height + margin.top + margin.bottom)
+        .attr("transform", `translate(${margin.left}, ${margin.top})`);
         
     g.value = svg.value.append('g')
-        .attr('class', 'g-els mf')
+        .attr('class', 'g-els sdf')
         .attr("transform", `translate(${margin.left}, ${margin.top})`);
 
     // set up chart elements
     setAxes();
-    addAxes();
-    addBrush();
-    addBars();
-}
 
-const addBars = () => {
-    // add box
-    const yearTotals = dataYears.value.map(year => {
-        if(year.data.length > 0){
-            return {
-                d: year.year,
-                v: year.data.reduce((acc, curr) => acc + curr)
-            };
-        } else {
-            return {
-                d: year.year,
-                v: 0
-            };
-        }
-    })
-    g.value.selectAll('.mf.bar')
-        .data(yearTotals)
-        .enter()
+    // add clip-path element - removing content outside the chart
+    const defs = g.value.append('defs');
+    defs.append('clipPath')
+        .attr('id', 'flow-duration-box-clip')
         .append('rect')
-        .attr('class', 'mf bar')
-        .attr('x', 0)
-        .attr('y', d => {
-            console.log(d.d)
-            return yScale.value(d.d)
-        })
-        .attr('width', d => {
-            console.log(xScale.value(0))
-            return xScale.value(d.v)
-        })
-        .attr('height', (d, idx) => yScale.value.bandwidth())
-        .attr('fill', 'steelblue')
-};
+        .attr('width', width)
+        .attr('height', height);
 
-const addBrush = () => {
-    brushVar.value = d3.brushY()
-        .extent([[0, 0], [width, height]])
-        .on("end", brushEnded)
-        .on("start", (ev) => {
-            if(ev.selection[0] === ev.selection[1]){
-                emit('year-range-selected', monthAbbrList[0], monthAbbrList[monthAbbrList.length - 1])
-            }
-        })
-    
-    brushEl.value = svg.value.append("g")
-        .call(brushVar.value)
-        .attr('transform', `translate(${margin.left}, ${margin.top})`)
+    addAxes();
+    addBoxPlots();
+
+    setTimeout(() => {
+        addBrush();
+    })
+
+    g.value.on('mousemove', mouseMoved);
+    g.value.on('mouseout', mouseOut);
 }
 
-const brushEnded = (event) => {
-    const selection = event.selection;
-    if (!event.sourceEvent || !selection) return;
+const mouseOut = () => {
+    showTooltip.value = false;
+}
 
-    const [y0, y1] = selection.map(d => {
-        return scaleBandInvert(yScale.value)(d)
-    });
+const mouseMoved = (event) => {
+    const [gX, gY] = d3.pointer(event, svg.value.node());
+    if (gX < margin.left || gX > width + margin.right) return;
+    if (gY > height + margin.top) return;
+    const date = scaleBandInvert(xScale.value)(gX - xScale.value.bandwidth());
+    const foundData = monthPercentiles.value.find(el => el.month === date);
 
-    brushedYearStart.value = y0;
-    brushedYearEnd.value = y1;
+    // some custom handling for the tooltip content, depending on their values
+    tooltipData.value = {};
+    tooltipData.value.Month = foundData.month
+    tooltipData.value.Max = foundData.max
+    tooltipData.value['75th %ile'] = foundData.p75
+    tooltipData.value.Median = foundData.p50
+    tooltipData.value['25th %ile'] = foundData.p25
+    tooltipData.value.Min = foundData.min
+    tooltipPosition.value = [event.pageX - 280, event.pageY - 100];
+    showTooltip.value = true;
+}
 
-    emit('year-range-selected', brushedYearStart.value, brushedYearEnd.value);
+const addBoxPlots = (scale = { x: xScale.value, y: yScale.value }) => {
+    monthPercentiles.value.forEach(month => {
+        // add maximum lines
+        g.value
+            .append('line')
+            .style('stroke', 'black')
+            .style('stroke-width', 2)
+            .attr('x1', scale.x(month.month))
+            .attr('y1', scale.y(month.max))
+            .attr('x2', scale.x(month.month) + scale.x.bandwidth())
+            .attr('y2', scale.y(month.max))
 
-    brushEl.value
-        .transition()
-        .ease(d3.easeLinear)
-        .call(
-            brushVar.value.move, 
-            [yScale.value(y0), yScale.value(y1)]
-        );
+        // add max to top of box line
+        g.value
+            .append('line')
+            .style('stroke', 'black')
+            .style("stroke-dasharray", "10, 3")
+            .style('stroke-width', 2)
+            .attr('x1', scale.x(month.month) + scale.x.bandwidth() / 2)
+            .attr('y1', scale.y(month.max))
+            .attr('x2', scale.x(month.month) + scale.x.bandwidth() / 2)
+            .attr('y2', scale.y(month.p75))
+
+        // add box
+        g.value
+            .append('rect')
+            .attr('x', scale.x(month.month))
+            .attr('y', scale.y(month.p75))
+            .attr('width', scale.x.bandwidth())
+            .attr('height', scale.y(month.p25) - scale.y(month.p75))
+            .attr('stroke', 'black')
+            .attr('fill', 'steelblue');
+        
+        // add median lines
+        g.value
+            .append('line')
+            .style('stroke', 'black')
+            .style('stroke-width', 2)
+            .attr('x1', scale.x(month.month))
+            .attr('y1', scale.y(month.p50))
+            .attr('x2', scale.x(month.month) + scale.x.bandwidth())
+            .attr('y2', scale.y(month.p50))
+
+        // add min to bottom of box line
+        g.value
+            .append('line')
+            .style('stroke', 'black')
+            .style("stroke-dasharray", "10, 3")
+            .style('stroke-width', 2)
+            .attr('x1', scale.x(month.month) + scale.x.bandwidth() / 2)
+            .attr('y1', scale.y(month.p25))
+            .attr('x2', scale.x(month.month) + scale.x.bandwidth() / 2)
+            .attr('y2', scale.y(month.min))
+            
+        // add minimum lines
+        g.value
+            .append('line')
+            .style('stroke', 'black')
+            .style('stroke-width', 2)
+            .attr('x1', scale.x(month.month))
+            .attr('y1', scale.y(month.min))
+            .attr('x2', scale.x(month.month) + scale.x.bandwidth())
+            .attr('y2', scale.y(month.min))
+            .attr('transform', `translate(0, 0)`)
+    })
 }
 
 const scaleBandInvert = (scale) => {
-    let domain = scale.domain().reverse();
+    let domain = scale.domain();
     var paddingOuter = scale(domain[0]);
     var eachBand = scale.step();
     return (val) => {
@@ -204,107 +245,169 @@ const scaleBandInvert = (scale) => {
     };
 };
 
+const addBrush = () => {
+    brushVar.value = d3.brushX()
+        .extent([[0, 0], [width, height]])
+        .on("end", brushEnded)
+        .on("start", (ev) => {
+            if(ev.selection[0] === ev.selection[1]){
+                emit('range-selected', monthAbbrList[0], monthAbbrList[monthAbbrList.length - 1])
+            }
+        })
+    
+    brushEl.value = g.value.append("g")
+        .call(brushVar.value);
+}
 
-const addAxes = () => {
+const brushEnded = (event) => {
+    const selection = event.selection;
+    if (!event.sourceEvent || !selection) return;
+    const [x0, x1] = selection.map(d => {
+        return scaleBandInvert(xScale.value)(d)
+    });
+
+    brushedStart.value = x0;
+    brushedEnd.value = x1;
+
+    emit('range-selected', brushedStart.value, brushedEnd.value);
+
+    brushEl.value
+        .transition()
+        .call(
+            brushVar.value.move, 
+            [xScale.value(x0), xScale.value(x1) + xScale.value.bandwidth()]
+        );
+}
+
+const addAxes = (scale = { x: xScale.value, y: yScale.value }) => {
     // x axis labels and lower axis line
     g.value.append('g')
         .attr('class', 'x axis')
-        .call(
-            d3.axisTop(xScale.value)
-            .ticks(5)
-            .tickFormat(d3.format(".1e"))
-    )
+        .call(d3.axisBottom(scale.x))
+        .attr('transform', `translate(0, ${height + 0})`)
+
+    g.value.append('text')
+        .attr('class', 'x axis-label')
+        .attr("transform", `translate(${width / 2}, ${height + 35})`)
+        .text('Date')
 
     // x axis labels and lower axis line
     g.value.append('g')
         .attr('class', 'y axis')
-        .call(
-            d3.axisLeft(yScale.value)
-        )
+        .call(d3.axisLeft(scale.y).ticks(5))
         .attr('transform', `translate(0, 0)`)
 
     g.value.append('text')
         .attr('class', 'y axis-label')
-        .attr("transform", `translate(-40, ${80})rotate(-90)`)
-        .text('Runoff (m³)')
+        .attr("transform", `translate(-40, ${height / 1.5})rotate(-90)`)
+        .text('Monthly Flow (m³/s)')
 }
 
 const setAxes = () => {
-    const total = dataYears.value.map(year => {
-        if(year.data.length > 0){
-            return year.data.reduce((acc, curr) => acc + curr)
-        } else {
-            return 0;
-        }
-    });
+    // set x-axis scale
+    xScale.value = d3.scaleBand()
+        .domain(monthAbbrList)
+        .range([0, width])
+        .padding(0.2)
 
     // set y-axis scale
-    xMax.value = d3.max(total);
-    xMax.value *= 1.10;
+    const valsToCheck = monthPercentiles.value.map(d => d);
 
-    // set x-axis scale
-    xScale.value = d3.scaleLinear()
-        .domain([0, xMax.value])
-        .range([0, width])
+    yMax.value = d3.max(valsToCheck.map(el => el.max));
+    yMax.value *= 1.10;
+    yMin.value = 0;
 
-    // Y axis displaying the years in reverse, descending
-    yScale.value = d3.scaleBand()
+    // Y axis
+    yScale.value = d3.scaleSymlog()
         .range([height, 0])
-        .domain(formattedChartData.value.map(el => new Date(el.d).getUTCFullYear()).reverse())
-        .paddingInner(0.3)
+        .domain([0, yMax.value]);
 }
 
-const processData = (rawData) => {
-    dataYears.value = [];
-    const start = new Date(rawData[0].d).getUTCFullYear();
-    const end = new Date(rawData[rawData.length - 1].d).getUTCFullYear();
+const processData = (data) => {
+    // sort data into month groups
+    sortDataIntoMonths(data);
+    monthPercentiles.value = [];
+    monthDataArr.value.forEach(month => {
+        monthPercentiles.value.push({
+            month: monthAbbrList[month.month],
+            max: percentile(month.data.filter(el => el.v !== null), 100),
+            p75: percentile(month.data.filter(el => el.v !== null), 75),
+            p50: percentile(month.data.filter(el => el.v !== null), 50),
+            p25: percentile(month.data.filter(el => el.v !== null), 25),
+            min: percentile(month.data.filter(el => el.v !== null), 0)
+        })
+    })
+}
 
-    for(let i = start; i <= end; i++){
-        dataYears.value.push({
-            year: i,
-            data: [], 
+const sortDataIntoMonths = (data) => {
+    monthAbbrList.forEach((_, idx) => {
+        const foundMonth = monthDataArr.value.find(el => el.month === idx);
+        const currMonthData = data.filter(el => {
+            return new Date(el.d).getMonth() === idx;
         });
-    }
-    
-    rawData.forEach(entry => {
-        const year = new Date(entry.d).getUTCFullYear();
-        const foundYear = dataYears.value.find(el => el.year === year);
-        if(!foundYear){
-            dataYears.value.push({
-                d: entry.d,
-                data: [ entry.v || 0.00 ]
+        if(!foundMonth){
+            monthDataArr.value.push({
+                month: idx,
+                data: currMonthData
             })
         } else {
-            foundYear.data.push(entry.v)
+            foundMonth.data = currMonthData
         }
+    })
+
+    monthDataArr.value.forEach(month => {
+        month.data.sort((a, b) => {
+            return a.v - b.v
+        });
     });
-    formattedChartData.value = rawData
 }
+
+const percentile = (sortedArray, p) => {
+    const index = (p / 100) * (sortedArray.length - 1);
+    const lower = Math.floor(index);
+    const upper = Math.ceil(index);
+
+    if (lower === upper) {
+        return sortedArray[lower].v;
+    }
+
+    const weight = index - lower;
+
+    return sortedArray[lower].v * (1 - weight) + sortedArray[upper].v * weight;
+} 
+
 </script>
 
 <style lang="scss">
-.date-selectors {
+
+// elements clipped by the clip-path rectangle
+.flow-duration-clipped {
+    clip-path: url('#flow-duration-box-clip');
+} 
+
+.flow-duration-container {
+    position: relative;
     display: flex;
-    align-items: center;
 
-    .selector {
-        width: 8rem;
-    }
-}
+    .flow-duration-tooltip {
+        position: absolute;
+        display: flex;
+        width: 10rem;
 
-.monthly-flow-container {
-    height: 100%;
-    max-height: 80vh;
-    overflow-y: auto;
+        .tooltip-header {
+            padding: 0.25rem;
+        }
 
-    #monthly-flow-chart-container {
-        height: 100%;
+        .tooltip-row {
+            padding: 0 0.7rem;
 
-        .svg-wrap-mf {
-            height: 100%;
-
-            .d3-chart-mf {
-                height: 100%;
+            &.box-val {
+                color: white;
+                background-color: steelblue;
+            }
+            &.val {
+                color: white;
+                background-color: rgb(41, 41, 41);
             }
         }
     }
