+++ client/src/components/streamflow/TotalRunoff.vue
@@ -1,33 +1,53 @@
 
 <template>
-    <h3>Flow Duration ({{ props.startEndMonths[0] }} - {{ props.startEndMonths[1] }})</h3>
-    <div id="total-runoff-chart-container">
-        <div class="svg-wrap-tr">
-            <svg class="d3-chart-tr">
-                <!-- d3 chart content renders here -->
-            </svg>
+    <h3>Total Runoff</h3>
+    <div class="date-selectors">
+        <q-select 
+            v-model="startYear"
+            class="selector"
+            label="Year From"
+            dense
+        />
+        <div class="q-mx-sm">
+            -
         </div>
+        <q-select 
+            v-model="endYear"
+            class="selector q-mx-sm"
+            label="Year to"
+            dense
+        />
+        <q-select 
+            v-model="specifiedMonth"
+            class="selector q-mx-sm"
+            label="Month"
+            dense
+        />
+        <q-btn 
+            class="text-bold q-mx-sm"
+            label="reset dates"
+            flat
+            color="primary"
+            @click="() => {}"
+        />
     </div>
-    <div 
-        v-if="showTooltip"
-        class="total-runoff-tooltip"
-        :style="`left: ${tooltipPosition[0]}px; top: ${tooltipPosition[1]}px`"
-    >
-        <q-card>
-            <div class="tooltip-header text-h6">
-                {{ tooltipData.exceedance }}% Exceedance Flow
+    <div class="monthly-flow-container">
+        <div id="monthly-flow-chart-container">
+            <div class="svg-wrap-mf">
+                <svg class="d3-chart-mf">
+                    <!-- d3 chart content renders here -->
+                </svg>
             </div>
-            <div class="tooltip-row">
-                {{ tooltipData.flow }} (m³/s)
-            </div>
-        </q-card>
+        </div>
     </div>
 </template>
 
 <script setup>
 import * as d3 from "d3";
 import { monthAbbrList } from '@/constants/dateHelpers.js';
-import { onMounted, ref, watch } from "vue";
+import { onMounted, ref, watch } from 'vue';
+
+const emit = defineEmits(['year-range-selected']);
 
 const props = defineProps({
     data: {
@@ -36,230 +56,275 @@ const props = defineProps({
     },
     startEndMonths: {
         type: Array,
-        required: true,
-    }
+        default: () => ['Jan', 'Dec'],
+    },
 });
 
+const monthDataArr = ref([]);
 const loading = ref(false);
+const monthPercentiles = ref([]);
 const formattedChartData = ref([]);
+const startYear = ref();
+const endYear = ref();
+const specifiedMonth = ref();
 
 // chart variables
 const svgWrap = ref();
 const svgEl = ref();
 const svg = ref();
 const g = ref();
-const yMax = ref(0);
-const yMin = ref(0);
-const flowLine = ref();
-const hoverCircle = ref();
-
-// chart scaling
 const xScale = ref();
 const yScale = ref();
+const xGrid = ref();
+const yGrid = ref();
+const xAxis = ref();
+const yAxis = ref();
+const xMax = ref();
+const yearRangeArr = ref([1914, 2022]);
+const dataYears = ref([1914]);
 
-// chart constants 
-const width = 600;
-const height = 300;
+// brush functionality
+const brushVar = ref();
+const brushEl = ref();
+const brushedYearStart = ref();
+const brushedYearEnd = ref();
+
+// chart constants
+const width = 400;
+const height = 1000;
 const margin = {
-    left: 50,
+    left: 60,
     right: 50,
-    top: 10,
+    top: 30,
     bottom: 50
 };
 
-// tooltip
-const showTooltip = ref(false);
-const tooltipData = ref();
-const tooltipPosition = ref();
-
 watch(() => props.startEndMonths, () => {
-    processData(props.data, props.startEndMonths);
-    initTotalRunoff();
-});
+    processData(props.data);
+    initializeMonthlyFlowChart();
+})
 
 onMounted(() => {
     loading.value = true;
-    processData(props.data, props.startEndMonths);
-    initTotalRunoff();
+    initializeMonthlyFlowChart();
     loading.value = false;
 });
 
-const initTotalRunoff = () => {
+const initializeMonthlyFlowChart = () => {
     if (svg.value) {
-        d3.selectAll('.g-els.tr').remove();
+        d3.selectAll('.g-els.mf').remove();
     }
-
-    svgWrap.value = document.querySelector('.svg-wrap-tr');
+    processData(props.data);
+    svgWrap.value = document.querySelector('.svg-wrap-mf');
     svgEl.value = svgWrap.value.querySelector('svg');
     svg.value = d3.select(svgEl.value)
         .attr("width", width + margin.left + margin.right)
-        .attr("height", height + margin.top + margin.bottom)
-        .attr("transform", `translate(${margin.left}, ${margin.top})`);
-
+        
     g.value = svg.value.append('g')
-        .attr('class', 'g-els tr')
+        .attr('class', 'g-els mf')
         .attr("transform", `translate(${margin.left}, ${margin.top})`);
 
-    // add clip-path element - removing content outside the chart
-    const defs = g.value.append('defs');
-    defs.append('clipPath')
-        .attr('id', 'total-runoff-box-clip')
-        .append('rect')
-        .attr('width', width)
-        .attr('height', height);
-
+    // set up chart elements
     setAxes();
     addAxes();
-    addFlowLine();
-
-    svg.value.on('mousemove', mouseMoved);
-    svg.value.on('mouseout', mouseOut);
-};
-
-const mouseOut = () => {
-    showTooltip.value = false;
+    addBrush();
+    addBars();
 }
 
-const mouseMoved = (event) => {
-    const [gX, gY] = d3.pointer(event, svg.value.node());
-    if (gX < margin.left || gX > width) return;
-    if (gY > height + margin.top) return;
-    const percentile = xScale.value.invert(gX);
-    const bisect = d3.bisector(d => d.exceedance).center;
-    const idx = bisect(formattedChartData.value, percentile - 10);
-    const data = formattedChartData.value[idx];
+const addBars = () => {
+    // add box
+    const yearTotals = formattedChartData.value.map(year => {
+        if(year.data.length > 0){
+            return {
+                d: year.year,
+                v: year.data.reduce((acc, curr) => acc + curr)
+            };
+        } else {
+            return {
+                d: year.year,
+                v: 0
+            };
+        }
+    })
 
-    addHoverCirlce(idx);
+    const bars = g.value.selectAll('.mf.bar')
+        .data(yearTotals)
+        .join('rect')
+        .attr('x', 0)
+        .attr('y', d => yScale.value(d.d))
+        .attr('width', 0)
+        .attr('height', () => yScale.value.bandwidth())
 
-    tooltipData.value = {
-        exceedance: data.exceedance ? data.exceedance.toFixed(2) : 0.00,
-        flow: data.value
-    };
-    tooltipPosition.value = [event.pageX - 280, event.pageY - 20];
-    showTooltip.value = true;
-};
-
-const addHoverCirlce = (index) => {
-    if(hoverCircle.value) g.value.selectAll('.dot').remove();
-    hoverCircle.value = g.value.append('circle')
-        .attr('class', 'dot')
-        .attr("r", 4)
-        .attr('cy', yScale.value(formattedChartData.value[index].value))
-        .attr('cx', xScale.value(formattedChartData.value[index].exceedance))
-        .attr('fill', 'darkblue')
-};
-
-const addFlowLine = () => {
-    flowLine.value = g.value.append('path')
-        .datum(formattedChartData.value)
-        .attr('fill', 'none')
-
-    flowLine.value
+    bars
         .transition()
         .duration(500)
-        .attr('stroke', 'steelblue')
-        .attr('stroke-width', 2)
-        .attr('class', 'sdf line median streamflow-clipped')
-        .attr('d', d3.line()
-            .x(d => xScale.value(d.exceedance))
-            .y(d => yScale.value(d.value))
-        )
-        .attr('transform', 'translate(1, 0)')
-        .delay(() => 100)
+        .attr('class', 'mf bar')
+        .attr('x', 0)
+        .attr('y', d => yScale.value(d.d))
+        .attr('width', d => xScale.value(d.v))
+        .attr('height', () => yScale.value.bandwidth())
+        .attr('fill', 'steelblue')
+};
+
+const addBrush = () => {
+    brushVar.value = d3.brushY()
+        .extent([[0, 0], [width, height]])
+        .on("end", brushEnded)
+        .on("start", (ev) => {
+            if(ev.selection[0] === ev.selection[1]){
+                emit('year-range-selected', dataYears.value[0].year, dataYears.value[dataYears.value.length - 1].year)
+            }
+        })
+    
+    brushEl.value = svg.value.append("g")
+        .call(brushVar.value)
+        .attr('transform', `translate(${margin.left}, ${margin.top})`)
 }
+
+const brushEnded = (event) => {
+    const selection = event.selection;
+    if (!event.sourceEvent || !selection) return;
+
+    const [y0, y1] = selection.map(d => {
+        return scaleBandInvert(yScale.value)(d)
+    });
+
+    brushedYearStart.value = y0;
+    brushedYearEnd.value = y1;
+
+    emit('year-range-selected', brushedYearStart.value, brushedYearEnd.value);
+
+    brushEl.value
+        .transition()
+        .ease(d3.easeLinear)
+        .call(
+            brushVar.value.move, 
+            [yScale.value(y0), yScale.value(y1)]
+        );
+}
+
+const scaleBandInvert = (scale) => {
+    let domain = scale.domain().reverse();
+    var paddingOuter = scale(domain[0]);
+    var eachBand = scale.step();
+    return (val) => {
+        var index = Math.floor((val - paddingOuter) / eachBand);
+        return domain[Math.max(0, Math.min(index, domain.length - 1))];
+    };
+};
+
 
 const addAxes = () => {
     // x axis labels and lower axis line
     g.value.append('g')
         .attr('class', 'x axis')
         .call(
-            d3.axisBottom(xScale.value)
-                // capture and remove the outermost 'padding' ticks
-                .tickFormat(d => d === -10 || d === 110 ? '' : `${d}%`)
-        )
-        .attr('transform', `translate(0, ${height + 0})`)
+            d3.axisTop(xScale.value)
+            .ticks(5)
+            .tickFormat(d3.format(".1e"))
+    )
 
     // x axis labels and lower axis line
     g.value.append('g')
         .attr('class', 'y axis')
         .call(
-                d3.axisLeft(yScale.value)
-                .ticks(3)
+            d3.axisLeft(yScale.value)
         )
         .attr('transform', `translate(0, 0)`)
 
     g.value.append('text')
         .attr('class', 'y axis-label')
-        .attr("transform", `translate(-40, ${height / 1.5})rotate(-90)`)
-        .text('Flow (m³/s)')
+        .attr("transform", `translate(-40, ${80})rotate(-90)`)
+        .text('Runoff (m³)')
 }
 
 const setAxes = () => {
-    // set x-axis scale
-    xScale.value = d3.scaleLinear()
-        .domain([-10, 110])
-        .range([0, width])
-
-    // set y-axis scale
-    yMax.value = d3.max(formattedChartData.value.map(el => el.value));
-    yMax.value *= 1.10;
-
-    // Y axis
-    yScale.value = d3.scaleSymlog()
-        .range([height, 0])
-        .domain([0, yMax.value]);
-}
-
-/**
- * handles the provided data and calculates the exceedance for the currently
- * selected range of data. 
- * 
- * @param dataToProcess - Array of all of the data returned by the API
- * @param range - start and end month array. eg. ['Jan', 'Dec']
- */
-const processData = (dataToProcess, range) => {
-    const start = monthAbbrList.indexOf(range[0])
-    const end = monthAbbrList.indexOf(range[1])
-
-    const dataInRange = dataToProcess.filter(el => {
-        const monthIdx = new Date(el.d).getMonth();
-        return monthIdx >= start && monthIdx <= end
-    })
-
-    formattedChartData.value = calculateExceedance(dataInRange.sort((a, b) => b.v - a.v))
-}
-
-const calculateExceedance = (sortedDescendingArray) => {
-    const N = sortedDescendingArray.length;
-    return sortedDescendingArray.map((value, i) => {
-        return {
-            value: value.v,
-            exceedance: ((i + 1) / N) * 100
+    const total = dataYears.value.map(year => {
+        if(year.data.length > 0){
+            return year.data.reduce((acc, curr) => acc + curr)
+        } else {
+            return 0;
         }
     });
+
+    // set y-axis scale
+    xMax.value = d3.max(total);
+
+    // set x-axis scale
+    xScale.value = d3.scaleLinear()
+        .domain([0, xMax.value])
+        .range([0, width])
+    
+    // Y axis displaying the years in reverse, descending
+    yScale.value = d3.scaleBand()
+        .range([height, 0])
+        .domain(formattedChartData.value.map(el => el.year).reverse())
+        .paddingInner(0.3)
+}
+
+const processData = (rawData) => {
+    dataYears.value = [];
+    const start = new Date(rawData[0].d).getUTCFullYear();
+    const end = new Date(rawData[rawData.length - 1].d).getUTCFullYear();
+
+    for(let i = start; i <= end; i++){
+        dataYears.value.push({
+            year: i,
+            data: [], 
+        });
+    }
+
+    const startMonthIdx = monthAbbrList.findIndex(el => el === props.startEndMonths[0]);
+    const endMonthIdx = monthAbbrList.findIndex(el => el === props.startEndMonths[1]);
+
+    rawData.forEach(entry => {
+        const year = new Date(entry.d).getUTCFullYear();
+        const foundYear = dataYears.value.find(el => el.year === year);
+
+        if(!foundYear){
+            dataYears.value.push({
+                d: new Date(entry.d),
+                data: [ entry.v ? entry.v : 0 ]
+            })
+        } else {
+            if(new Date(entry.d).getUTCMonth() >= startMonthIdx && new Date(entry.d).getUTCMonth() <= endMonthIdx){
+                foundYear.data.push(entry.v);
+            } else {
+                foundYear.data.push(0.00);
+            }
+
+        }
+    });
+
+    formattedChartData.value = dataYears.value;
 }
 </script>
 
 <style lang="scss">
-// elements clipped by the clip-path rectangle
-.total-runoff-clipped {
-    clip-path: url('#total-runoff-box-clip');
+.date-selectors {
+    display: flex;
+    align-items: center;
+
+    .selector {
+        width: 8rem;
+    }
 }
 
-.total-runoff-tooltip {
-        position: absolute;
-        display: flex;
-        flex-direction: column;
+.monthly-flow-container {
+    height: 100%;
+    max-height: 80vh;
+    overflow-y: auto;
 
-        .tooltip-header {
-            margin: 0 0.25rem;
-        }
+    #monthly-flow-chart-container {
+        height: 100%;
 
-        .tooltip-row {
-            margin: 0.25rem;
-            padding: 0 1rem;
-            color: white;
-            background-color: steelblue;
+        .svg-wrap-mf {
+            height: 100%;
+
+            .d3-chart-mf {
+                height: 100%;
+            }
         }
     }
+}
 </style>
